move to constants.go
